New resources created

+++ src/main/java/br/com/tiago/trixselection/service/imp/LocationServiceImp.java
@@ -7,14 +7,15 @@ import org.springframework.stereotype.Service;
 
 import br.com.tiago.trixselection.dao.LocationDao;
 import br.com.tiago.trixselection.model.Location;
+import br.com.tiago.trixselection.model.Tag;
 import br.com.tiago.trixselection.service.LocationService;
 
 @Service
-public class LocationServiceImp implements LocationService{
-	
+public class LocationServiceImp implements LocationService {
+
 	@Autowired
 	LocationDao locationDao;
-	
+
 	@Override
 	public List<Location> listAll() {
 		return locationDao.findAll();
@@ -26,29 +27,38 @@ public class LocationServiceImp implements LocationService{
 	}
 
 	@Override
-	public void create(Location location) {
+	public void create(Location location) throws Exception {
 		try {
 			locationDao.save(location);
 		} catch (Exception e) {
-			e.printStackTrace();
+			throw new Exception("Location's name already exists.");
 		}
 	}
 
 	@Override
-	public void update(Location location) {
-		try {
-			locationDao.update(location);
-		} catch (Exception e) {
-			e.printStackTrace();
-		}
+	public void update(Location location) throws Exception {
+		locationDao.update(location);
 	}
 
 	@Override
-	public void delete(Location location) {
-		try {
-			locationDao.delete(location);
-		} catch (Exception e) {
-			e.printStackTrace();
-		}		
+	public void delete(Location location) throws Exception {
+		if (location == null)
+			throw new Exception("Location not found.");
+		locationDao.delete(location);
+	}
+
+	@Override
+	public void setTagById(Tag tag, int locationId) throws Exception {
+		if (tag == null)
+			throw new Exception("The Tag don't exists");
+
+		Location location = getLocationById(locationId);
+
+		if (location == null)
+			throw new Exception("The Location not exists");
+
+		location.setTag(tag);
+
+		update(location);
 	}
 }
